+++ packages/ts-newskit/src/components/puzzles/puzzle-card/styles.ts
@@ -1,0 +1,12 @@
+import { styled, Block, getColorCssFromTheme } from 'newskit';
+import { NewsKitPuzzlePlaceholder } from './assets';
+
+export const Wrap = styled(Block)`
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+`;
+export const StyledNewsKitPuzzlePlaceholder = styled(NewsKitPuzzlePlaceholder)`
+  ${getColorCssFromTheme('color', 'puzzles010')};
+`;